refactor: add chapter table to bible database  - convert Short type to Integer  - deal with all side effect

+++ src/main/java/ch/hambak/lamp/bible/dto/BookResponse.java
@@ -27,5 +27,5 @@ public class BookResponse {
     private String abbrEng;
 
     @NotNull
-    private Short bookOrder;
+    private Integer bookOrder;
 }
